Add Hints, Abstract Components, Fix Styling Add hints to payment pages Abstract components from user page Fix styling across pages Use abstracted components
