+++ src/components/lib/Card.css
@@ -22,7 +22,7 @@
     display: block;
     font-size: 16px;
     color: var(--text-color);
-    margin-bottom: 25px;
+    margin-bottom: 15px;
     font-weight: bold;
 }
 
@@ -40,6 +40,22 @@
     border-color: var(--primary-color);
     box-shadow: 0 0 8px var(--primary-color);
     background-color: var(--input-bg-color-focused);
+}
+
+.card-content .input {
+    font-size: 14px;
+    border-radius: 4px;
+    border: 2px solid var(--input-border-color);
+    background-color: var(--input-bg-color);
+    color: var(--text-color);
+    padding: 10px;
+    margin-bottom: 25px;
+}
+
+.card-content .hint {
+    font-size: 13px;
+    color: var(--text-color);
+    margin-bottom: 5px;
 }
 
 .error {
