+++ src/components/user/Account.css
@@ -1,0 +1,7 @@
+.account-view {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    gap: 20px;
+}
