+++ src/components/user/Personal.css
@@ -1,0 +1,7 @@
+.personal-view {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    gap: 20px;
+}
